Updating hls_parallel_memcpy example code and cleanning the code
