+++ actions/hls_parallel_memcpy/hw/action_parallel_read_write.H
@@ -1,8 +1,8 @@
-#ifndef __ACTION_CREATE_VECTOR_H__
-#define __ACTION_CREATE_VECTOR_H__
+#ifndef __ACTION_PARALLEL_READ_WRITE_H__
+#define __ACTION_PARALLEL_READ_WRITE_H__
 
 /*
- * Copyright 2017 International Business Machines
+ * Copyright 2019 International Business Machines
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,11 +22,10 @@
 #include <ap_int.h>
 
 #include "hls_snap.H"
-#include <action_create_vector.h> /* Create Vector Job definition */
+#include <action_parallel_read_write.h> /* Parallel read/write Job definition */
 
 //--------------------------------------------------------------------
-// 22: simplify the data casting style
-#define RELEASE_LEVEL		0x00000022
+#define RELEASE_LEVEL		0x00000001
 
 typedef uint32_t mat_elmt_t;
 #define BURST_LENGTH 64
@@ -35,11 +34,11 @@ typedef mat_elmt_t word_t[DATA_PER_W];
 //---------------------------------------------------------------------
 // This is generic. Just adapt names for a new action
 // CONTROL is defined and handled by SNAP 
-// gpu_example_job_t is user defined in gpu_example/include/action_create_vector.h
+// parallel_memcpy_job_t is user defined in parallel_memcpy/include/action_parallel_read_write.h
 typedef struct {
 	CONTROL Control;	/*  16 bytes */
-	gpu_example_job_t Data;	/* up to 108 bytes */
-	uint8_t padding[SNAP_HLS_JOBSIZE - sizeof(gpu_example_job_t)];
+	parallel_memcpy_job_t Data;	/* up to 108 bytes */
+	uint8_t padding[SNAP_HLS_JOBSIZE - sizeof(parallel_memcpy_job_t)];
 } action_reg;
 
-#endif  /* __ACTION_CREATE_VECTOR_H__*/
+#endif  /* __ACTION_PARALLEL_READ_WRITE_H__*/
