+++ actions/hls_parallel_memcpy/sw/action_parallel_read_write.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 International Business Machines
+ * Copyright 2019 International Business Machines
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,9 @@
  */
 
 /*
- * Example to use the FPGA to generate a vetor of size vectorSize
+ * Example to read(write) data from(to) HOST in FPGA in parallel.
+ * Buffers are storded in a buffer for reading(writing) and swap between 
+ * each iteration.
  */
 
 #include <stdio.h>
@@ -33,7 +35,7 @@
 
 #include <snap_internal.h>
 #include <snap_tools.h>
-#include <action_create_vector.h>
+#include <action_parallel_read_write.h>
 
 static int mmio_write32(struct snap_card *card,
 			uint64_t offs, uint32_t data)
@@ -51,29 +53,69 @@ static int mmio_read32(struct snap_card *card,
 	return 0;
 }
 
-/* Main program of the software action */
+
+static void read_addr(uint8_t *flag, uint32_t **addr){
+	unsigned long address = 0x0ull;
+	for(int i = 0; i<(int)sizeof(uint64_t); i++){
+		address += (unsigned long)(flag[i+1] << 8*i);
+	}
+	
+	*addr = (uint32_t *)address;
+}
+
 static int action_main(struct snap_sim_action *action,
-		       void *job, unsigned int job_len)
+		void *job, unsigned int job_len)
 {
-	struct gpu_example_job *js = (struct gpu_example_job *)job;
-	uint32_t *dst;
-	uint64_t len;
-	size_t i;
-
-	/* No error checking ... */
-	act_trace("%s(%p, %p, %d)  type_out=%d jobsize %ld bytes\n",
-		  __func__, action, job, job_len, js->out.type,
-		  sizeof(*js));
-
-
+	printf("...");
+	struct parallel_memcpy_job *args = (struct parallel_memcpy_job *)job;
+	uint8_t *read_flag, *write_flag;
+	uint32_t *addr_read, *addr_write;
+	int max_iteration, vector_size;
+	
 	// get the parameters from the structure
-	len = js->vectorSize;
-	dst = (uint32_t *)(unsigned long)js->out.addr;
+	addr_read = (uint32_t *)(unsigned long)args->read.addr;
+    	addr_write = (uint32_t *)(unsigned long)args->write.addr;
+    	vector_size = args->vector_size;
+    	read_flag = (uint8_t *)(unsigned long)args->read_flag.addr;
+    	write_flag = (uint8_t *)(unsigned long)args->write_flag.addr;
+	max_iteration = args->max_iteration;
 
-	// software action processing : Create a vector of size vectorSize
-        for ( i = 0; i < len; i++ ) {
-                dst[i] = (uint32_t)i;
-        }
+	(void)job_len;
+	int i = 0;
+	size_t size = vector_size*sizeof(uint32_t);
+	uint32_t *buffer1 = malloc(size);
+	uint32_t *buffer2 = malloc(size);
+
+	buffer1[0] = 1;
+	buffer2[0] = 1;
+
+	printf("FPGA SW Action starts");
+	while (i<max_iteration) {
+		sleep(0.000001);
+		if ((read_flag[0] == 1) && (write_flag[0] == 1)){
+
+			//pointer switch
+			switch (i%2){
+				case 0:
+					memcpy(buffer1,addr_read,size);
+					memcpy(addr_write,buffer2,size);
+					break;
+				case 1:
+					memcpy(buffer2,addr_read,size);
+					memcpy(addr_write,buffer1,size);
+					break;
+			}
+
+			read_addr(read_flag,&addr_read);
+			read_addr(write_flag,&addr_write);
+			read_flag[0] = 0;
+			write_flag[0] = 0;
+			
+			i++;
+		}
+	}
+
+
 
 	// update the return code to the SNAP job manager
 	action->job.retc = SNAP_RETC_SUCCESS;
@@ -86,7 +128,7 @@ static int action_main(struct snap_sim_action *action,
 static struct snap_sim_action action = {
 	.vendor_id = SNAP_VENDOR_ID_ANY,
 	.device_id = SNAP_DEVICE_ID_ANY,
-	.action_type = GPU_EXAMPLE_ACTION_TYPE, // Adapt with your ACTION NAME
+	.action_type = PARALLEL_MEMCPY_ACTION_TYPE, // Adapt with your ACTION NAME
 
 	.job = { .retc = SNAP_RETC_FAILURE, },
 	.state = ACTION_IDLE,
